requirements.txt
# requirements.txt

# --- Framework e Servidor ---
fastapi==0.115.2
uvicorn[standard]==0.30.3
gunicorn==22.0.0

# --- Banco de Dados e ORM ---
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2

# --- Validação e Configuração ---
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
email-validator==2.2.0

# --- Segurança e Autenticação ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# --- Testes (opcional, mas bom ter para consistência) ---
pytest==8.3.2
httpx==0.27.0

# --- Geração de Dados de Mockup ---
Faker==26.0.0

# --- Dependências Principais (geralmente instaladas junto com as acima) ---
# É uma boa prática fixar as versões das dependências mais importantes
# para garantir builds consistentes.
starlette==0.37.2
anyio==4.4.0
h11==0.14.0